app/requirements.txt
azure-cosmos==4.3.0
azure-identity==1.12.0
dash==2.8.1
gunicorn==20.1.0
numpy==1.24.2
pandas==1.5.3
plotly==5.13.0
pyodbc==4.0.35
requests==2.28.2
opencensus-ext-azure==1.1.9
opencensus-ext-logging==0.1.1
dash-mantine-components==0.11.1
dash[diskcache] == 2.8.1
dash-iconify==0.1.2
dash-cytoscape==0.3.0
